Fix title of test
